common: generate markdown doc and use it for gh-pages

+++ utils/src2mans.sh
@@ -15,7 +15,9 @@ if [ "$DIR" == "" ]; then
 	exit 1
 fi
 
-mkdir -p html
+if which pandoc > /dev/null; then
+	mkdir -p md
+fi
 
 find $DIR -name '*.h' -print0 | while read -d $'\0' MAN
 do
@@ -33,12 +35,24 @@ do
 		cat $ERRORS
 	fi
 
-	for f in $(cat $MANUALS | xargs); do
-		# get rid of a FILE section (last two lines of the file)
-		mv $f $f.tmp
-		head -n -2 $f.tmp > $f
-		rm $f.tmp
-		cat $f | groff -mandoc -Thtml > html/$f.html
-	done
+	if [ -d md ]; then
+		for f in $(cat $MANUALS | xargs); do
+			# get rid of a FILE section (last two lines of the file)
+			mv $f $f.tmp
+			head -n -2 $f.tmp > $f
+			rm $f.tmp
+			
+			# generate a md file
+			pandoc -s $f -o $f.tmp1 -f man -t markdown
+			# remove the header 
+			tail -n +6 $f.tmp1 > $f.tmp2
+			# fix the name issue '**a **-' -> '**a** -'
+			sed -i '5s/ \*\*-/\*\* -/' $f.tmp2
+			# start with a custom header
+			cat ../../utils/mans_header.md > md/$f.md
+			cat $f.tmp2 >> md/$f.md
+			rm $f.tmp1 $f.tmp2
+		done
+	fi
 	rm $MANUALS $ERRORS
 done
